+++ src/app/features/amiibo/components/amiibo-item/amiibo-item.component.ts
@@ -1,9 +1,11 @@
-import { Component, input } from '@angular/core';
+import { Component, inject, input } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
 import { Amiibo } from '../../../../models/amiibo';
 
 import { AseButtonComponent } from '../../../../shared/ase-button/ase-button.component';
-import { RouterModule } from '@angular/router';
+
+import { CartService } from '../../../../services/cart.service';
 
 @Component({
   selector: 'amiibo-item',
@@ -14,9 +16,12 @@ import { RouterModule } from '@angular/router';
 })
 export class AmiiboItemComponent {
 
+  private _cartService = inject(CartService);
+
   public item = input.required<Amiibo>();
 
   public addToCart(): void {
     console.log("ADD TO CART");
+    this._cartService.addToCart(this.item());
   }
 }