+++ src/app/models/cart.ts
@@ -1,0 +1,11 @@
+import { Amiibo } from "./amiibo";
+
+export interface Cart {
+  items: Array<CartItem>;
+  total: number;
+}
+
+interface CartItem {
+  item: Amiibo;
+  quantity: number;
+}